backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
cloudinary==1.38.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
google-generativeai==0.3.2
